Intersphinx mapping and incremental build support (#19)
* Only call realpath on paths when using doxygen

Doxygen has issues with symbolic links, so we need to resolve these. Though for other tools, the symbolic link probably leads you outside the workspace, which creates trouble difficulties

* First version of INTERSPHINX_MAPPING

* (sphinx) Use full path of cwd, instead of dot

sphinx will resolve symbolic links, when resolving the dot. This is unwanted. Passing the cwd instead of the dot, still setting the cwd, eliminates the problem. So no resolving to outside the workspace.

* Use catkin_tools pluggable spaces

* Split deps from doc_deps

* Remove unused import

* Store Doxygen output dir instead of subfolder

* Set intersphinx_mapping in job_env

Might be good idea to remove intersphinx_mapping from job_env after the stage

* Cache used rosdoc config

* unset INTERSPHINX_MAPPING from job-env afterwards

* Add yaml dump funtionstage

* Use yaml dump function from this module

* Add write_file functionstage

* Use write_file function from this module

* Add unset_env functionstage

* Use unset_env function from this module

* sphinx>=4.3

* catkin_tools>=0.8.0

* sphinx>=4.3.0